requirements.txt
Django==2.2.4
gunicorn==19.9.0

# Heroku staging debug tools
django-debug-toolbar==1.11
django-querycount==0.7.0

# Database
dj-database-url==0.4.2
psycopg2==2.8.3

# Storage
whitenoise==3.3.0


# this is all the latest as of oct 19 2019
djangorestframework==3.10.3
djoser==2.0.3
django-axes==5.0.18
djangorestframework_simplejwt==4.3.0
